// frontend/src/app/models/user.ts
export class User {
  id?: number;
  email?: string;
  name?: string;
  password?: string;
  country?: string;
  city?: string;
  first_name?: string;
  last_name?: string;
  phone?: string;
  address?: string;
  company?: string;
  company_id?: number;
  website?: string;
  photo?: string;
  verified?: boolean;
}