// frontend/src/app/models/customer.ts
export class Customer {
  id?: number;
  owner?: number;
  first_name?: string;
  last_name?: string;
  email?: string;
  company?: string;
  phone?: string;
  address?: string;
  apartment?: string;
  city?: string;
  country?: string;
  region?: string;
  postal_code?: string;
  image?: string;
  socket_state: string;
}